refactor: move non-paging UserList overload to core:ui and use it in FavoritesScreen
